+++ src/service/yukicoder.rs
@@ -2,7 +2,7 @@ use crate::errors::{ScrapeError, ScrapeResult, ServiceErrorKind, ServiceResult};
 use crate::service::download::DownloadProgress;
 use crate::service::session::HttpSession;
 use crate::service::{
-    Contest, ExtractZip, LoginOutcome, PrintTargets as _, RetrieveLangsOutcome, RetrieveLangsProps,
+    Contest, ExtractZip, LoginOutcome, RetrieveLangsOutcome, RetrieveLangsProps,
     RetrieveTestCasesOutcome, RetrieveTestCasesOutcomeProblem, RetrieveTestCasesProps, Service,
     SessionProps, SubmitOutcome, SubmitProps, ZipEntries, ZipEntriesSorting,
 };
@@ -38,40 +38,37 @@ pub(crate) fn login(props: SessionProps, term: impl Term) -> ServiceResult<Login
 
 pub(crate) fn retrieve_testcases(
     props: (SessionProps, RetrieveTestCasesProps<String>),
-    mut term: impl Term,
+    term: impl Term,
 ) -> ServiceResult<RetrieveTestCasesOutcome> {
     let (sess_props, retrieve_props) = props;
     let retrieve_props = retrieve_props
         .convert_problems(CaseConversion::Upper)
         .parse_contest()
         .unwrap();
-    retrieve_props.print_targets(term.stderr())?;
     Yukicoder::try_new(sess_props, term)?.retrieve_testcases(&retrieve_props)
 }
 
 pub(crate) fn retrieve_langs(
     props: (SessionProps, RetrieveLangsProps<String>),
-    mut term: impl Term,
+    term: impl Term,
 ) -> ServiceResult<RetrieveLangsOutcome> {
     let (sess_props, retrieve_props) = props;
     let retrieve_props = retrieve_props
         .convert_problem(CaseConversion::Upper)
         .parse_contest()
         .unwrap();
-    retrieve_props.print_targets(term.stderr())?;
     Yukicoder::try_new(sess_props, term)?.retrieve_langs(retrieve_props)
 }
 
 pub(crate) fn submit(
     props: (SessionProps, SubmitProps<String>),
-    mut term: impl Term,
+    term: impl Term,
 ) -> ServiceResult<SubmitOutcome> {
     let (sess_props, submit_props) = props;
     let submit_props = submit_props
         .convert_problem(CaseConversion::Upper)
         .parse_contest()
         .unwrap();
-    submit_props.print_targets(term.stderr())?;
     Yukicoder::try_new(sess_props, term)?.submit(&submit_props)
 }